ref(installer): user account setup moves to authentication/accounts.py
* create_users, set_user_password, lock_root_account, add_to_seat_group keep
  thin Installer wrappers for the profiles and scripts that call them
* enable_sudo/enable_doas had no outside caller and go private with the move
* stash package list travels along, schema_gen follows
